refactor(config): Centralize path policy
